scripts/python/Motor-Control-UI/requirements.txt
PyQt5==5.15.11
pyserial==3.5
matplotlib==3.10.8
numpy==2.4.2
